requirements copy.txt
asgiref==3.4.1
attrs==21.2.0
backoff==1.10.0
certifi==2021.10.8
chardet==4.0.0
click==8.0.3
cookies==2.2.1
coverage==6.1.1
cycler==0.11.0
django-common-helpers==0.9.2
django-coverage-plugin==2.0.1
django-cron==0.5.1
django-extensions==3.1.3
django-json-widget==1.1.1
django==3.2.9
flake8==4.0.1
future==0.18.2
googleapis-common-protos==1.53.0
grpcio==1.41.1
idna==3.3
iniconfig==1.1.1
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
mypy==0.910
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.21.3
opentelemetry-api==1.6.2
opentelemetry-exporter-jaeger-proto-grpc==1.6.2
opentelemetry-exporter-jaeger-thrift==1.6.2
opentelemetry-exporter-jaeger==1.6.2
opentelemetry-exporter-otlp-proto-grpc==1.6.2
opentelemetry-exporter-otlp==1.6.2
opentelemetry-proto==1.6.2
opentelemetry-sdk==1.6.2
opentelemetry-semantic-conventions==0.25b2
packaging==21.2
Pillow==8.4.0
pip-upgrader==1.4.15
pluggy==1.0.0
protobuf==3.19.1
psycopg2==2.9.3
py==1.10.0
pycodestyle==2.8.0
pyparsing==2.4.7
pytest-django==4.4.0
pytest==6.2.5
python-logging-loki==0.3.1
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
thrift==0.15.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3


asgiref                            3.5.1
certifi                            2021.10.8
charset-normalizer                 2.0.12
Deprecated                         1.2.13
django-prometheus                  2.2.0
future                             0.18.2
idna                               3.3
mccabe                             0.6.1
opentelemetry-api                  1.11.1
opentelemetry-instrumentation      0.31b0
opentelemetry-instrumentation-wsgi 0.30b1
opentelemetry-semantic-conventions 0.30b1
opentelemetry-util-http            0.30b1
pip                                22.1.1
prometheus-client                  0.14.1
protobuf                           3.20.1
pycodestyle                        2.8.0
pyflakes                           2.4.0
setuptools                         58.1.0
six                                1.16.0
sqlparse                           0.4.2
thrift                             0.16.0
typing_extensions                  4.2.0
urllib3                            1.26.9
wrapt                              1.14.1